Basic Java Syntax
This is the first class that introduces the Java syntax. It shows how to
comment in Java ( the source itself is heavily commented ), its also
shows how to interact in the simplest of methods with the key. It source
contains how to variables, write to the console, concatenate two or more
strings, etc.