GPA-GradePointAverage- Calculator
This is second lecture. It introduces the concept of control statements
such as if-else, while loop, for loop, switch case, etc. It also briefly
describes introduces the concept of data structures with Array. The
focus of the class to learn how to use control statements to control the
flow of program.